Add files via upload
Sina CubeRotator is a graphical project. I first thought about how objects are seen and analyzed by human eyes and the brain. Then, I innovated mathematical equations to describe how the rotation of an object (like a cube) can affect its image on a monitor. Then, I presented an algorithm and developed it which shows a cube that is rotating in a 3D space. 
This is a philosophy project rather than a computer! Because I spent most of my time thinking instead of coding.
The executive file is located in the folder called "dist". (JavaRuntimeEnvironment must be installed on your computer for running this application.)