Add files via upload
Sina Snake Game is an interesting game designed and developed in both Java and C#. The game is published in two versions. The version called "Snake" includes versions in which the snake can move in a linear or rotary direction. Another version called SmartSnake is equipped with an intelligence algorithm that computer can lead the snake intelligently.
For Java Versions: The executive file is located in the folder called "dist". (JavaRuntimeEnvironment must be installed on your computer for running this application.)
For C# Versions: The executive file is located in the folder called "bin\debug" or "bin\release".