Add files via upload
TreeGame is an interesting game. In this game, player must stop the growth of the tree. The challenge is that, he only can cut the black branches, while all the branches change their colors each time. Also, the branch points are really small, so player must focus on the mouse movement.
The executive file is located in the folder called "dist". (JavaRuntimeEnvironment must be installed on your computer for running this application.)