Add files via upload
Tron is a game that can amuse you for hours. It consists of two phases. In the hard phase, the computer can control its movement intelligently, so the player has a low chance to win the game. 
The executive file is located in the folder called "dist". (JavaRuntimeEnvironment must be installed on your computer for running this application.)